requirements.txt
# Core dependencies
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# News scraping
feedparser==6.0.11
newspaper3k==0.2.8

# Database
sqlalchemy==2.0.25

# LLM
ollama==0.1.6

# Vector database (for project matching)
chromadb==0.4.22
sentence-transformers==2.3.1

# YouTube
google-api-python-client==2.111.0
youtube-transcript-api==0.6.2

# UI
streamlit==1.30.0
plotly==5.18.0

# Utilities
python-dotenv==1.0.1
pandas==2.1.4
tqdm==4.66.1
tenacity==8.2.3

# Development
pytest==7.4.4
black==23.12.1
flake8==7.0.0